backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
PyPDF2==3.0.1
pytesseract==0.3.10
Pillow==10.1.0
requests==2.31.0
pymysql==1.1.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
joblib==1.3.2
tensorflow==2.15.0
